scripts: hid_configurator: Move DFU to separate file
Change moves DFU related code to separate file.

Jira:DESK-593
